+++ src-tauri/src/desktop.rs
@@ -2330,6 +2330,7 @@ async fn authenticate_oauth(
     secrets::set_secret(&server_id, secrets::HTTP_AUTH_KEY, &res.access_token)?;
     remote::store_oauth_state(
         &server_id,
+        Some(res.issuer),
         &res.token_endpoint,
         &res.client_id,
         res.refresh_token,